Consolidate conversation state into stores/conversationStore
Merge selectedConversationStore + the conversation DI kit (state/actions/
use-conversation) + ConversationPanel's watch effect into one module-singleton
Solid store on the contacts pattern (#609). Selection now drives the message
watch directly; logout resets via the feature setup().

Delete the production-dead private/P2P transport seam (transportMode,
PrivateMessageTransport, P2PChatEnvelope, PrivateSessionSignaling,
CallChannelBridge, unreadCount, system messages) and the unused in-memory
adapter; ChatMessage drops source/delivery/actions.

Split the 920-line ConversationPanel into view-only components/
ConversationPanel + MessageList + Composer reading the store.
