+++ src/stores/conversationStore.test.js
@@ -1,0 +1,309 @@
+import { beforeEach, describe, expect, it, vi } from 'vite-plus/test';
+
+const mocks = vi.hoisted(() => ({
+  getLoggedInUserId: vi.fn(),
+  getLoggedInUserProfile: vi.fn(),
+  createD1MessageRepositoryFromEnv: vi.fn(),
+  markConversationRead: vi.fn(),
+  recordConversationActivity: vi.fn(),
+  resolveDirectConversationId: vi.fn(),
+  getContactById: vi.fn(),
+  cacheContactConversationId: vi.fn(),
+  uploadConversationFile: vi.fn(),
+  deleteConversationFile: vi.fn(),
+  sendMessageNotification: vi.fn(),
+  compressImage: vi.fn(),
+}));
+
+vi.mock('../auth/index.js', () => ({
+  getLoggedInUserId: mocks.getLoggedInUserId,
+}));
+vi.mock('./userProfileStore', () => ({
+  getLoggedInUserProfile: mocks.getLoggedInUserProfile,
+}));
+vi.mock('./message-repository', () => ({
+  createD1MessageRepositoryFromEnv: mocks.createD1MessageRepositoryFromEnv,
+}));
+vi.mock('./conversation-activity', () => ({
+  markConversationRead: mocks.markConversationRead,
+  recordConversationActivity: mocks.recordConversationActivity,
+}));
+vi.mock('./conversations-client', () => ({
+  resolveDirectConversationId: mocks.resolveDirectConversationId,
+}));
+vi.mock('./contactsStore.js', () => ({
+  getContactById: mocks.getContactById,
+  cacheContactConversationId: mocks.cacheContactConversationId,
+}));
+vi.mock('./filesStore.js', () => ({
+  uploadConversationFile: mocks.uploadConversationFile,
+  deleteConversationFile: mocks.deleteConversationFile,
+}));
+vi.mock('../features/push-notifications/index.js', () => ({
+  getPushNotifications: () => ({
+    sendMessageNotification: mocks.sendMessageNotification,
+  }),
+}));
+vi.mock('@lib/media/image-compress.js', () => ({
+  compressImage: mocks.compressImage,
+}));
+
+function envelope(overrides) {
+  return {
+    messageId: 'msg-1',
+    conversationId: 'conversation-1',
+    senderId: 'user-a',
+    sentAt: 1,
+    delivery: 'persistent',
+    reactions: [],
+    payload: { type: 'text', text: 'hello' },
+    ...overrides,
+  };
+}
+
+describe('conversationStore', () => {
+  /** Captured per test by the fake repository. */
+  let watch;
+  let repo;
+
+  beforeEach(() => {
+    vi.resetModules();
+    vi.clearAllMocks();
+
+    const storage = new Map();
+    vi.stubGlobal('localStorage', {
+      getItem: (key) => storage.get(String(key)) ?? null,
+      setItem: (key, value) => storage.set(String(key), String(value)),
+      removeItem: (key) => storage.delete(String(key)),
+    });
+
+    watch = { emit: null, error: null, unsubscribed: false };
+    repo = {
+      createMessageId: vi.fn(() => 'reserved-1'),
+      loadMessages: vi.fn(() => []),
+      watchRecentMessages: vi.fn((conversationId, onMessages, onError) => {
+        watch.emit = onMessages;
+        watch.error = onError;
+        return () => {
+          watch.unsubscribed = true;
+        };
+      }),
+      send: vi.fn(() => ({ id: 'reserved-1', sentAt: 10 })),
+      markConversationRead: vi.fn(),
+      setMyReaction: vi.fn(),
+    };
+    mocks.createD1MessageRepositoryFromEnv.mockReturnValue(repo);
+
+    mocks.getLoggedInUserId.mockReturnValue('me');
+    mocks.getLoggedInUserProfile.mockReturnValue({ displayName: 'Me' });
+    mocks.getContactById.mockReturnValue({
+      contactId: 'contact-1',
+      conversationId: 'conversation-1',
+    });
+  });
+
+  it('persists the selected contact id for direct opens', async () => {
+    const store = await import('./conversationStore.ts');
+
+    await store.openDirectConversation('contact-1', { displayUI: false });
+
+    expect(store.loadSelectedContactId()).toBe('contact-1');
+    expect(repo.watchRecentMessages).toHaveBeenCalledWith(
+      'conversation-1',
+      expect.any(Function),
+      expect.any(Function),
+    );
+  });
+
+  it('persists direct selections opened from raw selection state', async () => {
+    const store = await import('./conversationStore.ts');
+
+    store.open({
+      conversationId: 'conversation-1',
+      remoteParticipantIds: ['contact-2'],
+      displayUI: true,
+    });
+
+    expect(store.loadSelectedContactId()).toBe('contact-2');
+  });
+
+  it('merges watcher snapshots in chronological order and maps file envelopes', async () => {
+    const store = await import('./conversationStore.ts');
+    await store.openDirectConversation('contact-1');
+
+    watch.emit([
+      envelope({ messageId: 'msg-3', sentAt: 3 }),
+      envelope({ messageId: 'msg-1', sentAt: 1 }),
+      envelope({
+        messageId: 'msg-2',
+        sentAt: 2,
+        payload: {
+          type: 'file',
+          fileName: 'demo.png',
+          mimeType: 'image/png',
+          fileSize: 123,
+          storage: {
+            provider: 'r2',
+            bucket: 'hangvidu-files',
+            key: 'conversation-files/conversation-1/msg-2',
+          },
+        },
+      }),
+      envelope({
+        messageId: 'sys-1',
+        sentAt: 4,
+        payload: { type: 'system', systemType: 'noop' },
+      }),
+    ]);
+
+    const state = store.getConversationState();
+    expect(state.history).toBe('ready');
+    expect(state.messages.map((msg) => msg.id)).toEqual([
+      'msg-1',
+      'msg-2',
+      'msg-3',
+    ]);
+    expect(state.messages[1].attachment).toMatchObject({
+      type: 'file',
+      fileName: 'demo.png',
+      storage: { provider: 'r2', bucket: 'hangvidu-files' },
+    });
+  });
+
+  it('records DM activity for the latest watcher message', async () => {
+    const store = await import('./conversationStore.ts');
+    await store.openDirectConversation('contact-1');
+
+    watch.emit([envelope({ messageId: 'msg-1', sentAt: 5 })]);
+
+    expect(mocks.recordConversationActivity).toHaveBeenCalledWith(
+      'contact-1',
+      'conversation-1',
+      5,
+      'user-a',
+    );
+  });
+
+  it('sends the draft under a reserved persistent id and clears it', async () => {
+    const store = await import('./conversationStore.ts');
+    await store.openDirectConversation('contact-1');
+    store.setConversationDraft('hello');
+
+    const result = await store.sendMessage();
+
+    expect(result).toBe(true);
+    expect(repo.createMessageId).toHaveBeenCalledWith('conversation-1');
+    expect(repo.send).toHaveBeenCalledWith(
+      expect.objectContaining({
+        messageId: 'reserved-1',
+        conversationId: 'conversation-1',
+        senderId: 'me',
+        senderName: 'Me',
+        payload: { type: 'text', text: 'hello' },
+      }),
+    );
+    const state = store.getConversationState();
+    expect(state.draft).toBe('');
+    expect(state.messages.map((msg) => msg.id)).toEqual(['reserved-1']);
+    expect(state.messages[0].status).toBe('sent');
+    expect(mocks.sendMessageNotification).toHaveBeenCalledWith(
+      expect.objectContaining({
+        recipientIds: ['contact-1'],
+        conversationId: 'conversation-1',
+        messageText: 'hello',
+      }),
+    );
+  });
+
+  it('dedupes the watcher echo against the in-flight optimistic message', async () => {
+    const store = await import('./conversationStore.ts');
+    await store.openDirectConversation('contact-1');
+    store.setConversationDraft('hello');
+
+    let resolveSend;
+    repo.send.mockReturnValue(
+      new Promise((resolve) => {
+        resolveSend = resolve;
+      }),
+    );
+
+    const sending = store.sendMessage();
+    // Live echo lands before the send acknowledgement (same reserved id).
+    watch.emit([
+      envelope({ messageId: 'reserved-1', senderId: 'me', sentAt: 10 }),
+    ]);
+    resolveSend({ id: 'reserved-1', sentAt: 10 });
+    await sending;
+
+    const state = store.getConversationState();
+    expect(state.messages.map((msg) => msg.id)).toEqual(['reserved-1']);
+    expect(state.messages[0].status).toBe('sent');
+  });
+
+  it('marks the optimistic message failed when the send rejects', async () => {
+    const store = await import('./conversationStore.ts');
+    await store.openDirectConversation('contact-1');
+    store.setConversationDraft('hello');
+    repo.send.mockRejectedValue(new Error('offline'));
+
+    const result = await store.sendMessage();
+
+    expect(result).toBe(false);
+    const state = store.getConversationState();
+    expect(state.messages[0].status).toBe('failed');
+    expect(mocks.sendMessageNotification).not.toHaveBeenCalled();
+  });
+
+  it('isolates push notification failures from the send result', async () => {
+    const store = await import('./conversationStore.ts');
+    await store.openDirectConversation('contact-1');
+    store.setConversationDraft('hello');
+    mocks.sendMessageNotification.mockImplementationOnce(() => {
+      throw new Error('push unavailable');
+    });
+
+    await expect(store.sendMessage()).resolves.toBe(true);
+  });
+
+  it('persists drafts per conversation and restores them on reopen', async () => {
+    vi.useFakeTimers();
+    try {
+      const store = await import('./conversationStore.ts');
+      await store.openDirectConversation('contact-1');
+
+      store.setConversationDraft('unsent thought');
+      vi.advanceTimersByTime(300);
+
+      mocks.getContactById.mockReturnValue({
+        contactId: 'contact-2',
+        conversationId: 'conversation-2',
+      });
+      await store.openDirectConversation('contact-2');
+      expect(store.getConversationState().draft).toBe('');
+
+      mocks.getContactById.mockReturnValue({
+        contactId: 'contact-1',
+        conversationId: 'conversation-1',
+      });
+      await store.openDirectConversation('contact-1');
+      expect(store.getConversationState().draft).toBe('unsent thought');
+    } finally {
+      vi.useRealTimers();
+    }
+  });
+
+  it('reset stops the watch and clears all conversation state', async () => {
+    const store = await import('./conversationStore.ts');
+    await store.openDirectConversation('contact-1');
+    watch.emit([envelope()]);
+
+    store.resetConversationStore();
+
+    expect(watch.unsubscribed).toBe(true);
+    expect(store.selection()).toBeNull();
+    const state = store.getConversationState();
+    expect(state.conversationId).toBeNull();
+    expect(state.messages).toEqual([]);
+    expect(state.history).toBe('idle');
+  });
+});
